maintain: move server TLS funcs
These function will be modified in the next commit. Moving them now so that changes
are more obvious.
