+++ internal/server/server_test.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"context"
 	"crypto/tls"
-	"crypto/x509"
 	"encoding/json"
 	"io/ioutil"
 	"net/http"
@@ -20,10 +19,8 @@ import (
 	"github.com/rs/zerolog"
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
-	"gotest.tools/v3/golden"
 
 	"github.com/infrahq/infra/api"
-	"github.com/infrahq/infra/internal/cmd/types"
 	"github.com/infrahq/infra/internal/logging"
 	"github.com/infrahq/infra/internal/server/data"
 )
@@ -487,42 +484,3 @@ func TestServer_PersistSignupUser(t *testing.T) {
 	// retry the authenticated endpoint
 	checkAuthenticated()
 }
-
-func TestTLSConfigFromOptions(t *testing.T) {
-	storage := map[string]secrets.SecretStorage{
-		"plaintext": &secrets.PlainSecretProvider{},
-		"file":      &secrets.FileSecretProvider{},
-	}
-
-	ca := golden.Get(t, "pki/ca.crt")
-	t.Run("user provided certificate", func(t *testing.T) {
-		opts := TLSOptions{
-			CA:          types.StringOrFile(ca),
-			Certificate: types.StringOrFile(golden.Get(t, "pki/localhost.crt")),
-			PrivateKey:  "file:testdata/pki/localhost.key",
-		}
-		config, err := tlsConfigFromOptions(storage, t.TempDir(), opts)
-		assert.NilError(t, err)
-
-		srv := httptest.NewUnstartedServer(noopHandler)
-		srv.TLS = config
-		srv.StartTLS()
-		t.Cleanup(srv.Close)
-
-		roots := x509.NewCertPool()
-		roots.AppendCertsFromPEM(ca)
-		client := &http.Client{
-			Transport: &http.Transport{
-				TLSClientConfig: &tls.Config{RootCAs: roots, MinVersion: tls.VersionTLS12},
-			},
-		}
-
-		resp, err := client.Get(srv.URL)
-		assert.NilError(t, err)
-		assert.Equal(t, resp.StatusCode, http.StatusOK)
-	})
-}
-
-var noopHandler = http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-	w.WriteHeader(http.StatusOK)
-})
